Windows: extend some type conversion definitions
Add some more type conversion helpers for macros used in `CreateFileW`
which adds verbosity by explicitly casting the imported types to
`DWORD`.

+++ Sources/Foundation/NSURL.swift
@@ -927,9 +927,8 @@ extension NSURL {
 #if os(Windows)
         let hFile: HANDLE = absolutePath.withCString(encodedAs: UTF16.self) {
           CreateFileW($0, GENERIC_READ,
-                      DWORD(FILE_SHARE_READ | FILE_SHARE_WRITE | FILE_SHARE_DELETE),
-                      nil, DWORD(OPEN_EXISTING),
-                      DWORD(FILE_FLAG_BACKUP_SEMANTICS), nil)
+                      FILE_SHARE_READ | FILE_SHARE_WRITE | FILE_SHARE_DELETE,
+                      nil, OPEN_EXISTING, FILE_FLAG_BACKUP_SEMANTICS, nil)
         }
         guard hFile == INVALID_HANDLE_VALUE else {
           defer { CloseHandle(hFile) }
